importer renamed, and properties changed
