+++ client/coral-framework/actions/index.js
@@ -1,7 +1,7 @@
 import * as authActions from './auth';
 import * as assetActions from './asset';
 import * as notificationActions from './notification';
-import {actionsImporter} from '../helpers/importer';
+import {actionsImporter} from '../helpers/plugins';
 
 export default {
   authActions,
