+++ client/coral-framework/reducers/index.js
@@ -1,7 +1,7 @@
 import auth from './auth';
 import user from './user';
 import asset from './asset';
-import {reducersImporter} from '../helpers/importer';
+import {reducersImporter} from '../helpers/plugins';
 
 export default {
   auth,
